Refactor: DatabaseManager의 동작 방식을 전체적으로 수정 및 리팩토링

+++ JTFitCalendar/Service/DatabaseManager.swift
@@ -15,6 +15,15 @@ final class DatabaseManager {
 	let persistentContainer: NSPersistentContainer
 	let mainContext: NSManagedObjectContext
 	
+	/// Section에 대한 Rows들이 포함된 Dictionary
+	private var dateFitnessLogDictionary: [Date: [FitnessLogEntity]] = [:]
+	/// Section에 대한 Date들이 들어있는 Array
+	private var sectionDates: [Date] = []
+	
+	var sectionCount: Int {
+		return sectionDates.count
+	}
+	
 	private init() {
 		let persistentContainer = NSPersistentContainer(name: "CoreDataModel")
 		persistentContainer.loadPersistentStores { description, error in
@@ -24,6 +33,25 @@ final class DatabaseManager {
 		}
 		self.persistentContainer = persistentContainer
 		self.mainContext = persistentContainer.viewContext
+		
+		let request = FitnessLogEntity.fetchRequest()
+		let sortByDate = NSSortDescriptor(keyPath: \FitnessLogEntity.date, ascending: false)
+		request.sortDescriptors = [sortByDate]
+		
+		do {
+			let fitnessLogs = try mainContext.fetch(request)
+			fitnessLogs.forEach { fitnessLogEntity in
+				guard let date = fitnessLogEntity.date else { return }
+				if sectionDates.contains(date) == false {
+					sectionDates.append(date)
+					dateFitnessLogDictionary[date] = [fitnessLogEntity]
+				} else {
+					dateFitnessLogDictionary[date]?.append(fitnessLogEntity)
+				}
+			}
+		} catch {
+			fatalError("Database Fetch Error")
+		}
 	}
 	
 	func insertFitnessLog(
@@ -46,30 +74,24 @@ final class DatabaseManager {
 		newFitnessLog.consumedCalorie = consumedCalorie ?? 0.0
 		newFitnessLog.result = fitnessResult
 		
+		if sectionDates.contains(date) == false {
+			sectionDates.append(date)
+			sectionDates.sort(by: >)
+			dateFitnessLogDictionary[date] = [newFitnessLog]
+		} else {
+			dateFitnessLogDictionary[date]?.append(newFitnessLog)
+		}
+		
 		saveChanges()
 	}
 	
-	func fetchFitnessLogs(for date: Date? = nil) -> [FitnessLogEntity] {
-		let request = FitnessLogEntity.fetchRequest()
-		
-		if let date {
-			let beginDate = Calendar.current.startOfDay(for: date)
-			if let nextDate = Calendar.current.date(byAdding: .day, value: 1, to: beginDate) {
-				let predicate = NSPredicate(
-					format: "date >= %@ AND date < %@", beginDate as NSDate, nextDate as NSDate
-				)
-				request.predicate = predicate
-			}
-		}
-		
-		let sortByDate = NSSortDescriptor(keyPath: \FitnessLogEntity.date, ascending: false)
-		request.sortDescriptors = [sortByDate]
-		
-		do {
-			return try mainContext.fetch(request)
-		} catch {
-			return []
-		}
+	func fitnessLogs(for date: Date) -> [FitnessLogEntity]? {
+		return dateFitnessLogDictionary[date]
+	}
+	
+	func fitnessLog(for indexPath: IndexPath) -> FitnessLogEntity? {
+		let sectionDate = sectionDates[indexPath.section]
+		return fitnessLogs(for: sectionDate)?[indexPath.row]
 	}
 	
 	func updateFitnessLog(
@@ -95,7 +117,14 @@ final class DatabaseManager {
 		saveChanges()
 	}
 	
-	func delete(entity: NSManagedObject) {
+	func deleteRow(fitnessLog entity: FitnessLogEntity) {
+		guard let date = entity.date else { return }
+		guard let dictionaryValueIndex = dateFitnessLogDictionary[date]?.firstIndex(of: entity) else { return }
+		if dateFitnessLogDictionary[date]?.count == 1 {
+			dateFitnessLogDictionary[date] = nil
+		} else {
+			dateFitnessLogDictionary[date]?.remove(at: dictionaryValueIndex)
+		}
 		mainContext.delete(entity)
 		saveChanges()
 	}
@@ -110,6 +139,25 @@ final class DatabaseManager {
 					print(error.localizedDescription)
 				}
 			}
+		}
+	}
+	
+	func sectionDate(forSection section: Int) -> Date {
+		return sectionDates[section]
+	}
+	
+	func rowCount(forSection section: Int) -> Int {
+		return dateFitnessLogDictionary[sectionDate(forSection: section)]?.count ?? 0
+	}
+	
+	func rowCount(forDate date: Date) -> Int {
+		return dateFitnessLogDictionary[date]?.count ?? 0
+	}
+	
+	func deleteSection(forSection section: Int) {
+		let date = sectionDate(forSection: section)
+		if dateFitnessLogDictionary[date] == nil {
+			sectionDates.remove(at: section)
 		}
 	}
 }
